refactor: split session product read bands
Replace mixed durable product roots with explicit read and write bands across async query modules, query store composition, and repository accessors.
