+++ polylogue/storage/backends/queries/__init__.py
@@ -10,10 +10,13 @@ from polylogue.storage.backends.queries import (
     publications,
     raw,
     runs,
-    session_product_profile_queries,
+    session_product_profile_reads,
+    session_product_profile_writes,
+    session_product_query_support,
     session_product_summary_queries,
     session_product_thread_queries,
-    session_product_timeline_queries,
+    session_product_timeline_reads,
+    session_product_timeline_writes,
     stats,
 )
 
@@ -27,9 +30,12 @@ __all__ = [
     "publications",
     "raw",
     "runs",
-    "session_product_profile_queries",
+    "session_product_profile_reads",
+    "session_product_profile_writes",
+    "session_product_query_support",
     "session_product_summary_queries",
     "session_product_thread_queries",
-    "session_product_timeline_queries",
+    "session_product_timeline_reads",
+    "session_product_timeline_writes",
     "stats",
 ]
