+++ polylogue/storage/session_product_refresh.py
@@ -25,9 +25,11 @@ async def refresh_session_products_for_conversation_async(
     transaction_depth: int = 0,
 ) -> dict[str, int]:
     from polylogue.lib.threads import build_session_threads
-    from polylogue.storage.backends.queries.session_product_profile_queries import replace_session_profile
+    from polylogue.storage.backends.queries.session_product_profile_writes import (
+        replace_session_profile,
+    )
     from polylogue.storage.backends.queries.session_product_thread_queries import replace_work_thread
-    from polylogue.storage.backends.queries.session_product_timeline_queries import (
+    from polylogue.storage.backends.queries.session_product_timeline_writes import (
         replace_session_phases,
         replace_session_work_events,
     )
@@ -111,7 +113,7 @@ async def delete_session_products_for_conversation_async(
     *,
     transaction_depth: int = 0,
 ) -> dict[str, int]:
-    from polylogue.storage.backends.queries.session_product_timeline_queries import (
+    from polylogue.storage.backends.queries.session_product_timeline_writes import (
         replace_session_phases,
         replace_session_work_events,
     )
